Partial fix on error

+++ topobenchmarkx/transforms/feature_liftings/projection_sum.py
@@ -45,8 +45,8 @@ class ProjectionSum(torch_geometric.transforms.BaseTransform):
             if f"x_{elem}" not in data:
                 idx_to_project = 0 if elem == "hyperedges" else int(elem) - 1
                 data["x_" + elem] = torch.matmul(
-                    abs(data["incidence_" + elem].t().long()),
-                    data[f"x_{idx_to_project}"],
+                    abs(data["incidence_" + elem].t().float()),
+                    data[f"x_{idx_to_project}"].float(),
                 )
         return data
